+++ src/epiinf/distribs/TreeDensity.java
@@ -73,8 +73,8 @@ public class TreeDensity extends Distribution {
                 
                 // Incoporate probability of no effect on tree
                 if (revEventList.get(idx).type == EpidemicEvent.EventType.INFECTION && k>1) {
-                    int N = revStateList.get(idx).I;
-                    logP += Math.log(1.0 - k*(k-1)/((double)N*(N-1)));
+                    double N = revStateList.get(idx).I;
+                    logP += Math.log(1.0 - k*(k-1)/(N*(N-1)));
                 }
                 
                 idx += 1;
@@ -90,7 +90,7 @@ public class TreeDensity extends Distribution {
             if (treeEvent.type == TreeEventList.TreeEventType.SAMPLE)
                 k += 1; // Sample
             else {
-                int N = revStateList.get(idx).I;
+                double N = revStateList.get(idx).I;
                 logP += Math.log(k*(k-1)/((double)N*(N-1)));
                 k -= 1; // Coalescence
             }
